Restructure project and targets
