+++ src/plugin/DllMain.cpp
@@ -1,4 +1,4 @@
-#include "Plugin.h"
+#include "plugin/Plugin.h"
 
 #include <thread>
