+++ src/plugin/IComponent.h
